Fix compile warnings

+++ libs/graphics/bitmap_font_4x6.c
@@ -168,10 +168,10 @@ void graphicsDrawChar4x6(JsGraphics *gfx, short x1, short y1, char ch) {
   idx = (idx/5)*6;
   int y;
   for (y=0;y<6;y++) {
-    unsigned short line = LCD_FONT_4X6[idx + y] >> (cidx*3);
-    if (line&4) graphicsSetPixel(gfx, x1+0, y+y1, gfx->data.fgColor);
-    if (line&2) graphicsSetPixel(gfx, x1+1, y+y1, gfx->data.fgColor);
-    if (line&1) graphicsSetPixel(gfx, x1+2, y+y1, gfx->data.fgColor);
+    int line = LCD_FONT_4X6[idx + y] >> (cidx*3);
+    if (line&4) graphicsSetPixel(gfx, (short)(x1+0), (short)(y+y1), gfx->data.fgColor);
+    if (line&2) graphicsSetPixel(gfx, (short)(x1+1), (short)(y+y1), gfx->data.fgColor);
+    if (line&1) graphicsSetPixel(gfx, (short)(x1+2), (short)(y+y1), gfx->data.fgColor);
   }
 }
